Split out archive and envelope into separate mods
