+++ src/lib.rs
@@ -1,2 +1,3 @@
+pub mod archive;
+pub mod envelope;
 pub mod error;
-pub mod szdat;
